requirements.txt
openai
tavily-python
datasets
pandas
matplotlib
seaborn
tqdm
numpy
scikit-learn
statsmodels
python-dotenv
jupyter